Renamed Shader's header and source files to gl_shader.h and .cpp respectively to match the other opengl files in the project.
