+++ main/main.vcxproj
@@ -220,7 +220,7 @@
       <PrecompiledHeaderFile Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">../pch.h</PrecompiledHeaderFile>
       <PrecompiledHeaderFile Condition="'$(Configuration)|$(Platform)'=='Release|x64'">../pch.h</PrecompiledHeaderFile>
     </ClCompile>
-    <ClCompile Include="src\gl_engine\shader.cpp">
+    <ClCompile Include="src\gl_engine\gl_shader.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Use</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Use</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Use</PrecompiledHeader>
@@ -269,7 +269,7 @@
     <ClInclude Include="src\gl_engine\gl_camera.h" />
     <ClInclude Include="src\gl_engine\gl_mesh.h" />
     <ClInclude Include="src\gl_engine\gl_texture.h" />
-    <ClInclude Include="src\gl_engine\shader.h" />
+    <ClInclude Include="src\gl_engine\gl_shader.h" />
     <ClInclude Include="src\openglApp.h" />
     <ClInclude Include="src\pch.h" />
     <ClInclude Include="src\Renderer.h" />
