+++ main/src/gl_engine/gl_shader.cpp
@@ -1,5 +1,5 @@
 #include "../pch.h"
-#include "shader.h"
+#include "gl_shader.h"
 
 #include <errno.h>
 #include <string>
