+++ main/src/openglApp.h
@@ -5,7 +5,7 @@
 #include <string>
 #include <memory>
 
-#include "gl_engine/shader.h"
+#include "gl_engine/gl_shader.h"
 #include "gl_engine/gl_texture.h"
 #include "gl_engine/gl_mesh.h"
